backend/pyproject.toml
[project]
name = "backend-django"
version = "0.1.0"
description = "Django Backend"
requires-python = ">=3.13"
dependencies = [
    "Django>=5.2",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.3",
    "django-cors-headers>=4.4",
    "drf-yasg>=1.21",
    "daphne>=4.1",
    "channels>=4.1",
    "channels-redis>=4.2",
    "celery>=5.4",
    "django-redis>=5.4",
    "whitenoise>=6.7",
    "python-dotenv>=1.0",
    "python-json-logger>=2.0",
    "psycopg[binary]>=3.2",
    "redis>=5.0",
    "requests>=2.32",
    "pandas>=2.2",
    "numpy>=1.26",
    "openpyxl>=3.1",
    "pyyaml>=6.0",
    "PyJWT>=2.9",
    "pysmb>=1.2.10",
    "cryptography>=43.0",
    "gunicorn>=22.0",
]

[dependency-groups]
dev = ["ruff", "pytest", "pytest-django"]
test = ["pytest", "pytest-django"]
all = ["ruff", "pytest", "pytest-django"]

[tool.ruff]
line-length = 256
target-version = "py313"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.uv]
package = false
link-mode = "copy"

[tool.uv_scripts]
dev = "python manage.py runserver 0.0.0.0:8008"
start = "daphne -b 0.0.0.0 -p 8008 --application-close-timeout 30 backend.asgi:application"
format = "ruff format"
check = "ruff check"
test = "pytest -q"
lock = "uv lock --upgrade"
